+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Vulture` GitHub action
 - `Woke` GitHub action
 - Badges showing the status of GitHub actions on README page
+- File conversion endpoints accept optional parameters to set the following key/values: `submissionName, clinvarSubmissionReleaseStatus, assertionCriteria`
 ### Changed
 - New API schema introduced on 2022-11-21
 - Modified file parsing code to create submission jsons compliant with the new API